requirements.txt
<<<<<<< HEAD
requests>=2.31.0
PyYAML>=6.0.1
=======
requests>=2.31.0
PyYAML>=6.0.1
>>>>>>> 7f8c21ed401cad84e5bd773bf2ed8098230a339e
streamlit>=1.28.0